Missing error handling for Mage::getResourceSingleton
That global static class method does error handling not only by throwing
exceptions but also by returning FALSE.

This change fixes the flaw by introducing a type-check on the variable by
returning void on falsely.

How to reproduce:

When that model has been disabled in config.xml, there is no such
cleanLogs() method available as this is a non-object. See #30.